Refactor username generation to use a shared Trait and remove code duplication.
- Extracted optimized `generate_username` logic into `plugins/zeneyer-auth/includes/Auth/Trait-Username-Generator.php`.
- Updated `Google_Provider` and `Password_Auth` to use the `Username_Generator` trait.
- Removed redundant `stripos` check from the loop as the SQL query already ensures correct prefix filtering.
- Verified functionality and performance with updated benchmark script (2 queries for 1000 collisions).
