+++ plugins/zeneyer-auth/includes/Auth/class-password-auth.php
@@ -11,6 +11,7 @@ namespace ZenEyer\Auth\Auth;
 use WP_Error;
 
 class Password_Auth {
+    use Username_Generator;
     
     /**
      * Login with email and password
@@ -327,56 +328,4 @@ class Password_Auth {
         return true;
     }
     
-    /**
-     * Generate unique username from email
-     *
-     * @param string $email
-     * @return string
-     */
-    private static function generate_username($email) {
-        global $wpdb;
-
-        $username = sanitize_user(substr($email, 0, strpos($email, '@')));
-        
-        // Ensure uniqueness
-        if (!username_exists($username)) {
-            return $username;
-        }
-
-        // Optimization: Fetch all colliding usernames in one query instead of looping
-        // This prevents N+1 queries when many users share the same base name
-        $query = $wpdb->prepare(
-            "SELECT user_login FROM {$wpdb->users} WHERE LOWER(user_login) LIKE %s",
-            $wpdb->esc_like(strtolower($username)) . '%'
-        );
-
-        $taken_usernames = $wpdb->get_col($query);
-
-        $max_suffix = 0;
-        
-        foreach ($taken_usernames as $taken) {
-            // Check if it follows the pattern "username" + "number"
-            // We only care if it starts with our base username (case-insensitive)
-            if (stripos($taken, $username) !== 0) {
-                continue;
-            }
-
-            $suffix = substr($taken, strlen($username));
-
-            // If it's the base username itself
-            if (empty($suffix)) {
-                continue;
-            }
-
-            // If the suffix is numeric, track the max
-            if (ctype_digit($suffix)) {
-                $suffix_int = (int) $suffix;
-                if ($suffix_int > $max_suffix) {
-                    $max_suffix = $suffix_int;
-                }
-            }
-        }
-        
-        return $username . ($max_suffix + 1);
-    }
 }
